+++ vampire.h
@@ -8,7 +8,7 @@ class Vampire final: public Enemy {
     static bool isHostile;
     public:
         Vampire();
-        int dropGold() const override;
+        double dropGold() const override;
         void setHostile(bool isHostile) override ;
         bool getHostile() const override ;
         ~Vampire();
